SDIT-2148 Refactor audit to allow reuse
* Create subclass NomisEntity for audit columns with same name
* Create function to convert to DTO
* Improve audit person test

TODO Use for all person contact entities
