+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisprisonerapi/jpa/Person.kt
@@ -13,14 +13,12 @@ import jakarta.persistence.OneToMany
 import jakarta.persistence.SequenceGenerator
 import jakarta.persistence.Table
 import org.hibernate.Hibernate
-import org.hibernate.annotations.Generated
 import org.hibernate.annotations.JoinColumnOrFormula
 import org.hibernate.annotations.JoinColumnsOrFormulas
 import org.hibernate.annotations.JoinFormula
 import org.hibernate.annotations.SQLRestriction
 import org.hibernate.type.YesNoConverter
 import java.time.LocalDate
-import java.time.LocalDateTime
 
 @Entity
 @Table(name = "PERSONS")
@@ -138,14 +136,6 @@ data class Person(
   @Convert(converter = YesNoConverter::class)
   val keepBiometrics: Boolean = false,
 
-  @ManyToOne(fetch = LAZY)
-  @JoinColumn(name = "CREATE_USER_ID", insertable = false, updatable = false)
-  val createStaffUserAccount: StaffUserAccount? = null,
-
-  @ManyToOne(fetch = LAZY)
-  @JoinColumn(name = "MODIFY_USER_ID", insertable = false, updatable = false)
-  val modifyStaffUserAccount: StaffUserAccount? = null,
-
   /* columns not mapped
   OCCUPATION_CODE - always null
   CRIMINAL_HISTORY_TEXT - always null
@@ -167,51 +157,7 @@ data class Person(
   NAME_SEQUENCE - always null
    */
 
-) {
-  @Column(name = "CREATE_USER_ID", insertable = false, updatable = false)
-  @Generated
-  lateinit var createUsername: String
-
-  @Column(name = "CREATE_DATETIME", insertable = false, updatable = false)
-  @Generated
-  lateinit var createDatetime: LocalDateTime
-
-  @Column(name = "MODIFY_USER_ID", insertable = false, updatable = false)
-  @Generated
-  var modifyUserId: String? = null
-
-  @Column(name = "MODIFY_DATETIME", insertable = false, updatable = false)
-  @Generated
-  var modifyDatetime: LocalDateTime? = null
-
-  @Column(name = "AUDIT_TIMESTAMP", insertable = false, updatable = false)
-  @Generated
-  var auditTimestamp: LocalDateTime? = null
-
-  @Column(name = "AUDIT_USER_ID", insertable = false, updatable = false)
-  @Generated
-  var auditUserId: String? = null
-
-  @Column(name = "AUDIT_MODULE_NAME", insertable = false, updatable = false)
-  @Generated
-  var auditModuleName: String? = null
-
-  @Column(name = "AUDIT_CLIENT_USER_ID", insertable = false, updatable = false)
-  @Generated
-  var auditClientUserId: String? = null
-
-  @Column(name = "AUDIT_CLIENT_IP_ADDRESS", insertable = false, updatable = false)
-  @Generated
-  var auditClientIpAddress: String? = null
-
-  @Column(name = "AUDIT_CLIENT_WORKSTATION_NAME", insertable = false, updatable = false)
-  @Generated
-  var auditClientWorkstationName: String? = null
-
-  @Column(name = "AUDIT_ADDITIONAL_INFO", insertable = false, updatable = false)
-  @Generated
-  var auditAdditionalInfo: String? = null
-
+) : NomisEntity() {
   override fun equals(other: Any?): Boolean {
     if (this === other) return true
     if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
